requirements.txt
Django==4.0.2
django-environ==0.8.1
gunicorn==20.0.4
psycopg2==2.8.5
whitenoise==6.0.0
django-rosetta==0.9.8

# Cria uma tabela para cada model de linguagem diferente mas com referência a um.
django-parler==2.3
